Cleaned up all the warnings during the build. Whew.
git-svn-id: svn://10.0.50.55/legends/trunk@1016 67207eb2-9daa-044d-9a34-b8567af74898

+++ code/src/External/STB/STB/stb.h
@@ -5381,7 +5381,7 @@ FILE *  stb_fopen(char *filename, char *mode)
    #else
    {
       strcpy(temp_full+p, "stmpXXXXXX");
-      int fd = stb_mktemp(temp_full);
+      int fd = (int)stb_mktemp(temp_full);
       if (fd == -1) return NULL;
       f = fdopen(fd, mode);
       if (f == NULL) {
